# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NestPro PG & Hostel OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings/{buildingId}/floors:
    get:
      operationId: listFloors
      tags: [floors]
      summary: List floors in a building
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Floor"
    post:
      operationId: createFloor
      tags: [floors]
      summary: Create a floor
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
  /floors/{id}:
    patch:
      operationId: updateFloor
      tags: [floors]
      summary: Update a floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
    delete:
      operationId: deleteFloor
      tags: [floors]
      summary: Delete a floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Rooms ────────────────────────────────────────────────────
...
  /guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List guests
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, checked_out]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Check in a guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}/checkout:
    post:
      operationId: checkoutGuest
      tags: [guests]
      summary: Check out a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"

  # ─── Payments ────────────────────────────────────────────────
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      parameters:
        - name: propertyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/occupancy:
    get:
      operationId: getOccupancyBreakdown
      tags: [dashboard]
      summary: Bed occupancy grid by property
      parameters:
        - name: propertyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccupancyBreakdown"
  /dashboard/revenue:
    get:
      operationId: getRevenueStats
      tags: [dashboard]
      summary: Monthly revenue stats (last 6 months)
      parameters:
        - name: propertyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueMonth"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed across all properties
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"

  # ─── Activity Logs ────────────────────────────────────────────
...
components:
  schemas:
...
    Floor:
      type: object
      required: [id, buildingId, number, name, createdAt]
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        number:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    FloorInput:
      type: object
      required: [number, name]
      properties:
        number:
          type: integer
        name:
          type: string
    FloorUpdate:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string

    # ─── Room ────────────────────────────────────────────────────
...
    BedWithLocation:
      type: object
      required: [id, roomId, label, status, roomNumber, floorName, buildingName]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        label:
          type: string
        status:
          type: string
          enum: [available, occupied, maintenance, reserved]
        monthlyRent:
          type: ["number", "null"]
        roomNumber:
          type: string
        floorName:
          type: string
        buildingName:
          type: string
        guestName:
          type: ["string", "null"]
        guestId:
          type: ["integer", "null"]

    # ─── Guest ───────────────────────────────────────────────────
    Guest:
      type: object
      required: [id, name, phone, bedId, propertyId, checkInDate, status, monthlyRent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        aadhaar:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        hometown:
          type: ["string", "null"]
        bedId:
          type: integer
        propertyId:
          type: integer
        checkInDate:
          type: string
          format: date
        checkOutDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [active, checked_out]
        monthlyRent:
          type: number
        depositAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    GuestInput:
      type: object
      required: [name, phone, bedId, propertyId, checkInDate, monthlyRent]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        aadhaar:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        occupation:
          type: string
        hometown:
          type: string
        bedId:
          type: integer
        propertyId:
          type: integer
        checkInDate:
          type: string
          format: date
        monthlyRent:
          type: number
        depositAmount:
          type: number
        notes:
          type: string
    GuestUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        aadhaar:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        occupation:
          type: string
        hometown:
          type: string
        monthlyRent:
          type: number
        depositAmount:
          type: number
        notes:
          type: string
    CheckoutInput:
      type: object
      required: [checkOutDate]
      properties:
        checkOutDate:
          type: string
          format: date
        notes:
          type: string

    # ─── Payment ─────────────────────────────────────────────────
...
    DashboardSummary:
      type: object
      required: [totalProperties, totalBeds, occupiedBeds, availableBeds, occupancyRate, totalGuests, activeGuests, totalRevenue, collectedRevenue, pendingRevenue, openComplaints, overduePayments]
      properties:
        totalProperties:
          type: integer
        totalBeds:
          type: integer
        occupiedBeds:
          type: integer
        availableBeds:
          type: integer
        occupancyRate:
          type: number
        totalGuests:
          type: integer
        activeGuests:
          type: integer
        checkInsToday:
          type: integer
        checkOutsToday:
          type: integer
        totalRevenue:
          type: number
        collectedRevenue:
          type: number
        pendingRevenue:
          type: number
        openComplaints:
          type: integer
        overduePayments:
          type: integer
    OccupancyBreakdown:
      type: object
      required: [propertyId, propertyName, buildings]
      properties:
        propertyId:
          type: integer
        propertyName:
          type: string
        buildings:
          type: array
          items:
            $ref: "#/components/schemas/BuildingOccupancy"
    BuildingOccupancy:
      type: object
      required: [buildingId, buildingName, floors]
      properties:
        buildingId:
          type: integer
        buildingName:
          type: string
        floors:
          type: array
          items:
            $ref: "#/components/schemas/FloorOccupancy"
    FloorOccupancy:
      type: object
      required: [floorId, floorName, rooms]
      properties:
        floorId:
          type: integer
        floorName:
          type: string
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/RoomOccupancy"
    RoomOccupancy:
      type: object
      required: [roomId, roomNumber, roomType, beds]
      properties:
        roomId:
          type: integer
        roomNumber:
          type: string
        roomType:
          type: string
        beds:
          type: array
          items:
            $ref: "#/components/schemas/BedWithLocation"
    RevenueMonth:
      type: object
      required: [month, year, label, collected, pending, total]
      properties:
        month:
          type: integer
        year:
          type: integer
        label:
          type: string
        collected:
          type: number
        pending:
          type: number
        total:
          type: number

    # ─── Activity ─────────────────────────────────────────────────
    ActivityLog:
      type: object
      required: [id, action, entity, description, createdAt]
      properties:
        id:
          type: integer
        action:
          type: string
        entity:
          type: string
        entityId:
          type: ["integer", "null"]
        description:
          type: string
        propertyId:
          type: ["integer", "null"]
        propertyName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time